Building-a-Real-Time-eCommerce-Churn-Early-Warning-System-with-Flask-Kafka-and-Next.js/kafka/requirements.txt
kafka-python
pandas
numpy
python-dotenv
requests
pymysql
sqlalchemy
redis
schedule
